+++ src/index.ts
@@ -2,7 +2,6 @@ import Flagship from './class/flagship/flagship';
 import { FlagshipNodeSdk } from './types';
 
 export const flagship: FlagshipNodeSdk = {
-    initSdk: (envId, config) => new Flagship(envId, config), // deprecated
     start: (envId, config) => new Flagship(envId, config)
 };